Update Controller, Service, Helper

+++ game/src/main/java/com/game/controller/PlayersController.java
@@ -5,12 +5,15 @@ import com.game.entity.Race;
 import com.game.model.Player;
 import com.game.service.PlayerService;
 import com.game.service.ServiceFilter;
+import com.game.service.ServiceHelper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.domain.Specification;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -18,7 +21,7 @@ import java.util.List;
 
 @RestController
 @RequestMapping("/rest/players")
-public class PlayersController {
+public class PlayersController extends ServiceHelper {
 
     @Autowired
     private PlayerService playerService;
@@ -46,7 +49,7 @@ public class PlayersController {
                                         @RequestParam(value = "pageNumber", required = false, defaultValue = "0") Integer pageNumber,
                                         @RequestParam(value = "pageSize", required = false, defaultValue = "3") Integer pageSize) {
         Pageable pageable = PageRequest.of(pageNumber, pageSize, Sort.by(order.getFieldName()));
-        List<Player> playerList = playerService.getFilteredList(Specification.where
+        return playerService.getFilteredList(Specification.where
                         (ServiceFilter.filterByName(name).and
                         (ServiceFilter.filterByTitle(title)).and
                         (ServiceFilter.filterByRace(race)).and
@@ -58,18 +61,22 @@ public class PlayersController {
                         (ServiceFilter.filterByLevelMoreThan(minLevel)).and
                         (ServiceFilter.filterByLevelLessThan(maxLevel)).and
                         (ServiceFilter.filterIsBanned(banned))), pageable).getContent();
-        return playerList;
     }
 
     //получать игрока по id;
     @GetMapping("/{id}")
-    public Player getPlayer(@PathVariable("id") String id) {
-        return playerService.getPlayer(id);
+    public Player getPlayer(@PathVariable Long id) {
+        if (!isValidID(id)) throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        try {
+            return playerService.getPlayer(id);
+        } catch (Exception e) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
     }
 
     //получать количество игроков, которые соответствуют фильтрам;
     @GetMapping("/count")
-    public Integer getPlayersCount (@RequestParam(value = "name", required = false) String name,
+    public Long getPlayersCount (@RequestParam(value = "name", required = false) String name,
                                    @RequestParam(value = "title", required = false) String title,
                                    @RequestParam(value = "race", required = false) Race race,
                                    @RequestParam(value = "profession", required = false) Profession profession,
@@ -80,28 +87,59 @@ public class PlayersController {
                                    @RequestParam(value = "maxExperience", required = false) Integer maxExperience,
                                    @RequestParam(value = "minLevel", required = false) Integer minLevel,
                                    @RequestParam(value = "maxLevel", required = false) Integer maxLevel) {
-        return playerService.getPlayersCount(name, title, race, profession, after, before, banned, minExperience, maxExperience, minLevel, maxLevel);
+        return playerService.getPlayersCount(Specification.where
+                        (ServiceFilter.filterByName(name).and
+                        (ServiceFilter.filterByTitle(title)).and
+                        (ServiceFilter.filterByRace(race)).and
+                        (ServiceFilter.filterByProfession(profession)).and
+                        (ServiceFilter.filterByBirthdayMoreThan(after)).and
+                        (ServiceFilter.filterByBirthdayLessThan(before)).and
+                        (ServiceFilter.filterByExpMoreThan(minExperience)).and
+                        (ServiceFilter.filterByExpLessThan(maxExperience)).and
+                        (ServiceFilter.filterByLevelMoreThan(minLevel)).and
+                        (ServiceFilter.filterByLevelLessThan(maxLevel)).and
+                        (ServiceFilter.filterIsBanned(banned))));
     }
 
     //создавать нового игрока;
     @PostMapping()
     public Player createPlayer (@RequestBody Player player) {
+        if (!isValidID(player.getId()) || !isValidName(player.getName()) || !isValidTitle(player.getTitle()) ||
+        !isValidExp(player.getExperience()) || !isValidDate(player.getBirthday())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        }
+        if (player.isBanned() == null) player.setBanned(false);
         return playerService.createPlayer(player);
     }
 
     //редактировать характеристики существующего игрока;
     @PostMapping("/{id}")
-    public Player updatePlayer(@PathVariable("id") String id, @RequestBody Player player) {
-        Player playerOld = playerService.getPlayer(id);
-        return playerService.updatePlayer(player, playerOld);
+    public Player updatePlayer(@PathVariable Long id, @RequestBody Player player) {
+        Player tempPlayer = getPlayer(id);
+        if (!isValidName(player.getName()) || !isValidTitle(player.getTitle()) ||
+                !isValidExp(player.getExperience()) || !isValidDate(player.getBirthday())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        }
+        if (player.getName() == null) player.setName(tempPlayer.getName());
+        if (player.getTitle() == null) player.setTitle(tempPlayer.getTitle());
+        if (player.getRace() == null) player.setRace(tempPlayer.getRace());
+        if (player.getProfession() == null) player.setProfession(tempPlayer.getProfession());
+        if (player.getExperience() == null) player.setExperience(tempPlayer.getExperience());
+        if (player.isBanned() == null) player.setBanned(tempPlayer.isBanned());
+        if (player.getBirthday() == null) player.setBirthday(tempPlayer.getBirthday());
+        setCurrentLevel(player);
+        setExpToTheNextLevel(player);
+        return playerService.updatePlayer(player);
     }
 
     //удалять игрока;
     @DeleteMapping("/{id}")
-    public void deletePlayer(@PathVariable("id") String id) {
-        playerService.deletePlayer(id);
+    public void deletePlayer(@PathVariable Long id) {
+        if (!isValidID(id)) throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        try {
+            playerService.deletePlayer(id);
+        } catch (Exception e) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
     }
-
-
-
 }
